HTTP signed requests (client side) (#594)
