+++ tests/perfclient.py
