+++ tests/requirements.txt
@@ -6,3 +6,4 @@ loguru
 coincurve
 psutil
 cimetrics>=0.2.1
+requests-http-signature
