+++ tests/schema.py
@@ -11,7 +11,6 @@ import shutil
 import random
 import infra.ccf
 import infra.proc
-import infra.jsonrpc
 import e2e_args
 
 from loguru import logger as LOG
